# Avoid copy when a string is matched in the chunk buffered iterator.