see https://github.com/golang/go/issues/25484.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>

```
benchcmp before.txt after.txt
benchmark                     old ns/op     new ns/op     delta
BenchmarkRead/none-16         81061549      46772599      -42.30%
BenchmarkRead/gzip-16         234061520     188522019     -19.46%
BenchmarkRead/lz4-16          106841476     72388064      -32.25%
BenchmarkRead/lz4-256k-16     103634343     72696346      -29.85%
BenchmarkRead/lz4-1M-16       103995402     72552087      -30.24%
BenchmarkRead/lz4-4M-16       104292989     75668354      -27.45%
BenchmarkRead/snappy-16       106606110     77175866      -27.61%

benchmark                     old allocs     new allocs     delta
BenchmarkRead/none-16         491005         2651           -99.46%
BenchmarkRead/gzip-16         491732         2784           -99.43%
BenchmarkRead/lz4-16          491055         2652           -99.46%
BenchmarkRead/lz4-256k-16     491050         2653           -99.46%
BenchmarkRead/lz4-1M-16       491074         2652           -99.46%
BenchmarkRead/lz4-4M-16       491053         2653           -99.46%
BenchmarkRead/snappy-16       491041         2652           -99.46%

benchmark                     old bytes     new bytes     delta
BenchmarkRead/none-16         134491948     183704        -99.86%
BenchmarkRead/gzip-16         134624025     195672        -99.85%
BenchmarkRead/lz4-16          134618181     192355        -99.86%
BenchmarkRead/lz4-256k-16     135055340     250397        -99.81%
BenchmarkRead/lz4-1M-16       136022749     324037        -99.76%
BenchmarkRead/lz4-4M-16       138728808     743780        -99.46%
BenchmarkRead/snappy-16       134584271     193373        -99.86%
```

Am I dreaming ?