@cyriltovena 2020-01-24
Alright, let's close this, seems dangerous since the string become mutable.